requirements.txt
python-dotenv==1.0.1

## LLM
openai==2.8.1
streamlit==1.51.0
tiktoken==0.12.0
pymupdf-layout==1.26.6
pymupdf4llm==0.1.9

## LANGCHAIN
langchain-openai==1.1.0
langchain-text-splitters==1.0.0
langchain==1.1.0

## POSTGRES
pgvector==0.3.1
psycopg2-binary==2.9.9
pydantic==2.12.4

## STREAMLIT
streamlit==1.51.0
streamlit-authenticator==0.4.2